requirements.txt
yt_dlp
pynacl
discord.py
asyncio
python-dotenv